#145 solve duplicate spec problem by moving the test cases
